Delete Confirmation

+++ src/app/report/report.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from "@angular/core";
 import { HttpService } from "../http.service";
+import { Observable, of } from 'rxjs';
 import { Router } from "@angular/router";
 import Swal from "sweetalert2";
 
@@ -22,12 +23,12 @@ export class ReportComponent implements OnInit {
     await this.http.getUserInfo(this.googleId).subscribe(data => {
       this.user = data
     })
-    this.http.getReports().subscribe(data => {
+    await this.http.getReports().subscribe(data => {
       for (var key in data) {
         if (data[key].user_id === this.user.id) {
-          this.userReports.push(data[key])
+          this.userReports.unshift(data[key])
         } else {
-          this.reports.push(data[key]);
+          this.reports.unshift(data[key]);
         }
       }
     });
@@ -35,18 +36,35 @@ export class ReportComponent implements OnInit {
 
   deleteReport(event, report) {
     console.log(report);
-    for (let i = 0; i < this.userReports.length; i++) {
-      if (this.userReports[i].id === report.id) {
-        this.userReports.splice(i, 1);
-      }
-    };
-    this.http.deleteReport(report.id).subscribe((response) => {
-      console.log(response);
-      if (response === "DELETED") {
-        Swal.fire("Report Deleted!", "success");
-        this.router.navigate([""]);
-      } else {
-        Swal.fire("Something went wrong...");
+    Swal.fire({
+      title: 'Are you sure?',
+      text: "You won't be able to revert this!",
+      type: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Yes, delete it!'
+    }).then((result) => {
+      if (result.value) {
+        for (let i = this.userReports.length - 1; i >= 0; i--) {
+          if (this.userReports[i].id === report.id) {
+            this.userReports.splice(i, 1);
+          }
+        };
+        this.http.deleteReport(report.id).subscribe((response) => {
+          console.log(response);
+          if (response === "DELETED") {
+            Swal.fire("Report Deleted!", "success");
+            this.router.navigate([""]);
+          } else {
+            Swal.fire("Something went wrong...");
+          }
+        })
+        Swal.fire(
+          'Deleted!',
+          'Your file has been deleted.',
+          'success'
+        )
       }
     })
   }
